refactor: split core types by domain behind a barrel
