+++ packages/core/src/types/mission.ts
@@ -1,0 +1,246 @@
+/**
+ * Mission domain — mission lifecycle, reports, flow control (quality gates,
+ * checkpoints, delays, approval gates), SLA, quality metrics, and scheduling.
+ */
+
+import type { TaskOutcome } from "./task.js";
+import type { ScopedNotificationRules } from "./notifications.js";
+
+// === Mission ===
+
+export type MissionStatus = "draft" | "scheduled" | "recurring" | "active" | "paused" | "completed" | "failed" | "cancelled";
+
+export interface Mission {
+  id: string;
+  name: string;         // "mission-1", "mission-2", or custom name
+  data: string;         // JSON mission content (tasks, team, etc.)
+  prompt?: string;      // original user prompt that generated this mission
+  status: MissionStatus;
+  /** Absolute deadline for the entire mission (ISO timestamp). */
+  deadline?: string;
+  /** Cron expression or ISO timestamp for scheduled execution. */
+  schedule?: string;
+  /** End date for recurring schedules (ISO timestamp). After this date the schedule stops firing and the mission transitions to completed. */
+  endDate?: string;
+  /** Minimum average score for the mission to be considered successful. */
+  qualityThreshold?: number;
+  /** Scoped notification rules — override or extend global rules for tasks in this mission. */
+  notifications?: ScopedNotificationRules;
+  /** How many times this mission has been executed. Incremented on each run (useful for recurring missions). */
+  executionCount?: number;
+  /**
+   * Opaque end-user identifier (OpenAI-compat `user`). Tasks generated by this
+   * mission inherit this value unless explicitly overridden at task level.
+   */
+  user?: string;
+  createdAt: string;
+  updatedAt: string;
+}
+
+/** Completion report for a mission — aggregated results across all tasks. */
+export interface MissionReport {
+  missionId: string;
+  group: string;
+  allPassed: boolean;
+  totalDuration: number;           // ms, sum of all task durations
+  tasks: {
+    title: string;
+    status: "done" | "failed";
+    duration: number;              // ms
+    score?: number;                // global assessment score (1-5)
+    filesCreated: string[];
+    filesEdited: string[];
+    outcomes?: TaskOutcome[];      // outcomes produced by this task
+  }[];
+  filesCreated: string[];          // aggregated across all tasks
+  filesEdited: string[];           // aggregated across all tasks
+  outcomes?: TaskOutcome[];        // aggregated outcomes across all tasks
+  avgScore?: number;               // average assessment score
+}
+
+// === Approval Gates ===
+
+export type ApprovalGateHandler = "auto" | "human";
+
+export interface ApprovalGateCondition {
+  /** JS-like expression evaluated against the hook payload.
+   *  For "auto" gates — if condition passes, task proceeds. If it fails, task is blocked.
+   *  For "human" gates — condition determines WHEN to trigger the gate. */
+  expression: string;
+}
+
+export interface ApprovalGate {
+  /** Unique gate ID. */
+  id: string;
+  /** Human-readable name. */
+  name: string;
+  /** "auto" = system evaluates condition. "human" = blocks for human approval. */
+  handler: ApprovalGateHandler;
+  /** Which lifecycle hook triggers this gate. */
+  hook: string;
+  /** Optional condition — when to activate the gate. */
+  condition?: ApprovalGateCondition;
+  /** Notification channels to alert on gate activation (for "human" gates). */
+  notifyChannels?: string[];
+  /** Timeout in ms (for "human" gates). 0 = no timeout. */
+  timeoutMs?: number;
+  /** Action when timeout expires. Default: "reject". */
+  timeoutAction?: "approve" | "reject";
+  /** Priority within the same hook point. Lower = first. Default: 100. */
+  priority?: number;
+  /** Max revision rounds before only approve/reject is allowed. Default: 3. */
+  maxRevisions?: number;
+  /** Include task outcomes as attachments in the approval notification. */
+  includeOutcomes?: boolean;
+}
+
+export type ApprovalStatus = "pending" | "approved" | "rejected" | "timeout";
+
+export interface ApprovalRequest {
+  /** Unique request ID. */
+  id: string;
+  /** Gate that triggered this request. */
+  gateId: string;
+  /** Gate name (denormalized for display). */
+  gateName: string;
+  /** Related task ID, if applicable. */
+  taskId?: string;
+  /** Related mission ID, if applicable. */
+  missionId?: string;
+  /** Current status. */
+  status: ApprovalStatus;
+  /** Hook payload snapshot at time of request. */
+  payload: unknown;
+  /** When the request was created. */
+  requestedAt: string;
+  /** When the request was resolved (approved/rejected/timeout). */
+  resolvedAt?: string;
+  /** Who resolved it (user ID, "system", "timeout"). */
+  resolvedBy?: string;
+  /** Optional resolution note. */
+  note?: string;
+}
+
+// === Quality Layer ===
+
+/** Quality gate defined within a mission — checkpoint between task phases. */
+export interface MissionQualityGate {
+  /** Gate name. */
+  name: string;
+  /** Tasks that must be completed before this gate is evaluated. */
+  afterTasks: string[];
+  /** Tasks that are blocked until this gate passes. */
+  blocksTasks: string[];
+  /** Minimum average score of `afterTasks` to pass. */
+  minScore?: number;
+  /** All `afterTasks` must have status "done" (not just completed — they must pass). */
+  requireAllPassed?: boolean;
+  /** Custom condition expression evaluated against gate context. */
+  condition?: string;
+  /** Notification channels to alert on gate pass/fail. */
+  notifyChannels?: string[];
+}
+
+/** Checkpoint defined within a mission — planned stopping point for human review.
+ *
+ * Unlike approval gates (which ask yes/no and auto-resume on approval),
+ * checkpoints unconditionally pause the mission until explicitly resumed.
+ * Use checkpoints for human-in-the-loop review at defined milestones. */
+export interface MissionCheckpoint {
+  /** Checkpoint name (used in events and notifications). */
+  name: string;
+  /** Tasks that must be completed before this checkpoint triggers. */
+  afterTasks: string[];
+  /** Tasks that are blocked until the checkpoint is resumed. */
+  blocksTasks: string[];
+  /** Notification channels to alert when the checkpoint is reached. */
+  notifyChannels?: string[];
+  /** Optional message included in the notification when the checkpoint triggers. */
+  message?: string;
+}
+
+/** Delay defined within a mission — timed wait between task groups.
+ *
+ * Unlike checkpoints (which pause until a human resumes), delays
+ * automatically resume after a specified duration elapses.
+ * The timer starts when ALL afterTasks reach a terminal state (done/failed).
+ * Use delays for cooldown periods, staggered rollouts, rate-limiting, etc. */
+export interface MissionDelay {
+  /** Delay name (used in events and notifications). */
+  name: string;
+  /** Tasks that must be completed before this delay timer starts. */
+  afterTasks: string[];
+  /** Tasks that are blocked until the delay timer expires. */
+  blocksTasks: string[];
+  /** ISO 8601 duration (e.g. "PT2H" = 2 hours, "PT30M" = 30 minutes, "P1D" = 1 day). */
+  duration: string;
+  /** Notification channels to alert when the delay starts / expires. */
+  notifyChannels?: string[];
+  /** Optional message included in the notification when the delay starts. */
+  message?: string;
+}
+
+/** SLA configuration for deadline monitoring. */
+export interface SLAConfig {
+  /** Percentage of deadline elapsed before emitting a warning (0-1). Default: 0.8 */
+  warningThreshold?: number;
+  /** Check interval in ms. Default: 30000 (30s). */
+  checkIntervalMs?: number;
+  /** Notification channels for SLA warnings. */
+  warningChannels?: string[];
+  /** Notification channels for SLA violations. */
+  violationChannels?: string[];
+  /** Action on SLA violation: "notify" (default) or "fail" (force-fail the task). */
+  violationAction?: "notify" | "fail";
+}
+
+/** Quality metrics snapshot for a single entity (task, agent, mission). */
+export interface QualityMetrics {
+  /** Entity identifier. */
+  entityId: string;
+  /** Entity type. */
+  entityType: "task" | "agent" | "mission";
+  /** Total assessments run. */
+  totalAssessments: number;
+  /** Assessments that passed. */
+  passedAssessments: number;
+  /** Average global score (1-5). */
+  avgScore?: number;
+  /** Minimum score observed. */
+  minScore?: number;
+  /** Maximum score observed. */
+  maxScore?: number;
+  /** Per-dimension average scores. */
+  dimensionScores: Record<string, number>;
+  /** Total retries consumed. */
+  totalRetries: number;
+  /** Total fix attempts consumed. */
+  totalFixes: number;
+  /** Deadlines met vs missed. */
+  deadlinesMet: number;
+  deadlinesMissed: number;
+  /** Last updated. */
+  updatedAt: string;
+}
+
+/** Scheduled mission entry — runtime artifact derived from Mission fields. */
+export interface ScheduleEntry {
+  /** Unique schedule ID. */
+  id: string;
+  /** Mission ID to execute. */
+  missionId: string;
+  /** Cron expression (e.g. "0 2 * * *") or ISO timestamp for one-shot. */
+  expression: string;
+  /** Whether this schedule recurs (derived from mission status === "recurring"). */
+  recurring: boolean;
+  /** Whether this schedule is active. */
+  enabled: boolean;
+  /** Last execution time (ISO). */
+  lastRunAt?: string;
+  /** Next scheduled execution time (ISO). */
+  nextRunAt?: string;
+  /** Deadline offset — auto-set task/mission deadline to N ms after execution start. */
+  deadlineOffsetMs?: number;
+  /** Created at. */
+  createdAt: string;
+}
